pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "yolo-object-detection"
version = "2.0.0"
description = "YOLOv8-Large Semiconductor Wafer Defect Detection with Triton Serving"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.1",
    "numpy>=1.24",
    "pandas>=2.0",
    "opencv-python-headless>=4.8",
    "pillow>=10.0",
    "matplotlib>=3.8",
    "mlflow>=2.10",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.6",
    "httpx>=0.26",
    "redis>=5.0",
    "onnx>=1.15",
    "onnxruntime>=1.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "black>=24.0",
    "locust>=2.20",
]
gpu = [
    "tritonclient[all]>=2.42",
    "onnxruntime-gpu>=1.17",
    "tensorrt>=8.6",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
